Move code from GameHandler.hpp to GameHandler.cpp Now GameHandler.hpp contains interface only.
